Refactor BattlePage
