+++ src/App.js
@@ -1,7 +1,7 @@
 import React,{Component} from 'react';
 import './App.css';
 import { Route, BrowserRouter as Router, Switch } from 'react-router-dom';
-import BattlePage from './components/BattlePage'
+import BattlePage from './components/BattlePage/BattlePage'
 import LoserPage from './components/Loser/LoserPage.js';
