+++ src/components/BattlePage/BattlePage.css
@@ -1,0 +1,10 @@
+.BattlePage {
+    text-align: center;
+    flex-wrap: wrap;
+    background-color: blue;
+    display:flex;
+    background: linear-gradient(217deg, rgba(255,0,0,.8), rgba(255,0,0,0) 70.71%), linear-gradient(127deg, rgba(0,255,0,.8), rgba(0,255,0,0) 70.71%), linear-gradient(336deg, rgba(0,0,255,.8), rgba(0,0,255,0) 70.71%);background: linear-gradient(217deg, rgba(255,0,0,.8), rgba(255,0,0,0) 70.71%), linear-gradient(127deg, rgba(0,255,0,.8), rgba(0,255,0,0) 70.71%), linear-gradient(336deg, rgba(0,0,255,.8), rgba(0,0,255,0) 70.71%);  display:flex;
+    flex-wrap: wrap;
+    justify-content: space-around;
+    height: 100vh;
+}
